[sk] Refactored print handler into logger
