+++ mage_ai/shared/logger.py
@@ -1,3 +1,4 @@
+from contextlib import contextmanager
 import logging
 import time
 
@@ -42,3 +43,22 @@ class VerboseFunctionExec:
                 print('DONE')
             else:
                 print('FAILED')
+
+
+class VerbosePrintHandler:
+    def __init__(self, start_msg, verbose=False):
+        self.verbose = verbose
+        if verbose:
+            print(start_msg)
+        self.exists_previous_message = False
+
+    @contextmanager
+    def print_msg(self, msg):
+        if self.verbose:
+            if self.exists_previous_message:
+                print(f'\r├─ ')
+            print(f'└─ {msg}...', end='')
+        yield msg
+        if self.verbose:
+            print('DONE', end='')
+            self.exists_previous_message = True
